requirements.txt
Flask~=3.0.3
pyaml~=24.7.0
pandas~=2.2.2
Flask-SQLAlchemy~=3.1.1
mariadb~=1.1.10